+++ src/service_identity/pyopenssl.py
@@ -6,14 +6,16 @@ from __future__ import absolute_import, division, print_function
 
 import warnings
 
+import six
+
 from pyasn1.codec.der.decoder import decode
 from pyasn1.type.char import IA5String
 from pyasn1.type.univ import ObjectIdentifier
 from pyasn1_modules.rfc2459 import GeneralNames
 
 from ._common import (
-    DNS_ID, CertificateError, DNSPattern, SRVPattern, URIPattern,
-    verify_service_identity
+    DNS_ID, CertificateError, DNSPattern, IPAddress_ID, IPAddressPattern,
+    SRVPattern, URIPattern, verify_service_identity
 )
 from .exceptions import SubjectAltNameWarning
 
@@ -46,6 +48,31 @@ def verify_hostname(connection, hostname):
     )
 
 
+def verify_ip_address(connection, ip_address):
+    """
+    Verify whether the certificate of *connection* is valid for *ip_address*.
+
+    :param OpenSSL.SSL.Connection connection: A pyOpenSSL connection object.
+    :param unicode ip_address: The IP address that *connection* should be
+        connected to.  Can be an IPv4 or IPv6 address.
+
+    :raises service_identity.VerificationError: If *connection* does not
+        provide a certificate that is valid for *ip_address*.
+    :raises service_identity.CertificateError: If the certificate chain of
+        *connection* contains a certificate that contains invalid/unexpected
+        data.
+
+    :returns: ``None``
+
+    .. versionadded:: 18.1.0
+    """
+    verify_service_identity(
+        cert_patterns=extract_ids(connection.get_peer_certificate()),
+        obligatory_ids=[IPAddress_ID(ip_address)],
+        optional_ids=[],
+    )
+
+
 ID_ON_DNS_SRV = ObjectIdentifier('1.3.6.1.5.5.7.8.7')  # id_on_dnsSRV
 
 
@@ -61,7 +88,7 @@ def extract_ids(cert):
     :return: List of IDs.
     """
     ids = []
-    for i in range(cert.get_extension_count()):
+    for i in six.moves.range(cert.get_extension_count()):
         ext = cert.get_extension(i)
         if ext.get_short_name() == b"subjectAltName":
             names, _ = decode(ext.get_data(), asn1Spec=GeneralNames())
@@ -69,6 +96,10 @@ def extract_ids(cert):
                 name_string = n.getName()
                 if name_string == "dNSName":
                     ids.append(DNSPattern(n.getComponent().asOctets()))
+                elif name_string == "iPAddress":
+                    ids.append(IPAddressPattern.from_bytes(
+                        n.getComponent().asOctets()
+                    ))
                 elif name_string == "uniformResourceIdentifier":
                     ids.append(URIPattern(n.getComponent().asOctets()))
                 elif name_string == "otherName":
@@ -82,6 +113,10 @@ def extract_ids(cert):
                             raise CertificateError(
                                 "Unexpected certificate content."
                             )
+                    else:  # pragma: nocover
+                        pass
+                else:  # pragma: nocover
+                    pass
 
     if not ids:
         # http://tools.ietf.org/search/rfc6125#section-6.4.4
@@ -100,6 +135,7 @@ def extract_ids(cert):
             "removed by major browsers and deprecated by RFC 2818.  "
             "service_identity will remove the support for it in mid-2018."
             .format(cn.decode("utf-8")),
-            SubjectAltNameWarning
+            SubjectAltNameWarning,
+            stacklevel=2,
         )
     return ids